Fixed flag takover communication
The server now has a correct value in the database for all inRange
sections of flags. The client only sends the server a message when he
either entered a flag radius or is outside one while still registered.
Also changed tooLittlePlayers check to only give a warning (allows users
to test the plugin on their own).